:recycle: Separate munch animation from state
